[update] split the file decode.rs
